refactor:change folders name
